@@ -10,24 +10,26 @@
     "format": "prettier --plugin-search-dir . --write ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^3.3.1",
-    "@sveltejs/kit": "^2.37.0",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/adapter-auto": "^6.1.0",
+    "@sveltejs/kit": "^2.42.2",
+    "@sveltejs/vite-plugin-svelte": "^6.2.0",
+    "@tailwindcss/postcss": "^4.1.13",
     "autoprefixer": "^10.4.21",
-    "eslint": "^8.57.1",
-    "eslint-config-prettier": "^8.10.2",
-    "eslint-plugin-svelte": "^2.46.1",
+    "eslint": "^9.35.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-svelte": "^3.12.3",
     "postcss": "^8.5.6",
-    "prettier": "^2.8.8",
-    "prettier-plugin-svelte": "^2.10.1",
-    "svelte": "^4.2.20",
-    "tailwindcss": "^3.4.17",
-    "vite": "^5.4.19"
+    "prettier": "^3.6.2",
+    "prettier-plugin-svelte": "^3.4.0",
+    "svelte": "^5.39.2",
+    "tailwindcss": "^4.1.13",
+    "vite": "^7.1.6"
   },
   "type": "module",
   "dependencies": {
-    "@iconify/svelte": "^4.0.2",
+    "@iconify/svelte": "^5.0.2",
     "@tailwindcss/typography": "^0.5.16",
+    "cross-spawn": "^7.0.6",
     "mdsvex": "^0.12.6"
   }
 }
