@@ -1,29 +1,29 @@
 <!-- This file is generated by @sveltejs/kit — do not edit it! -->
-
+<svelte:options runes={true} />
 <script>
-	import { setContext, afterUpdate, onMount, tick } from 'svelte';
+	import { setContext, onMount, tick } from 'svelte';
 	import { browser } from '$app/environment';
 
 	// stores
-	export let stores;
-	export let page;
-	
-	export let constructors;
-	export let components = [];
-	export let form;
-	export let data_0 = null;
-	export let data_1 = null;
+	let { stores, page, constructors, components = [], form, data_0 = null, data_1 = null } = $props();
 
 	if (!browser) {
 		setContext('__svelte__', stores);
 	}
 
-	$: stores.page.set(page);
-	afterUpdate(stores.page.notify);
+	if (browser) {
+		$effect.pre(() => stores.page.set(page));
+	} else {
+		stores.page.set(page);
+	}
+	$effect(() => {
+		stores;page;constructors;components;form;data_0;data_1;
+		stores.page.notify();
+	});
 
-	let mounted = false;
-	let navigated = false;
-	let title = null;
+	let mounted = $state(false);
+	let navigated = $state(false);
+	let title = $state(null);
 
 	onMount(() => {
 		const unsubscribe = stores.page.subscribe(() => {
@@ -39,16 +39,21 @@
 		return unsubscribe;
 	});
 
-	
+	const Pyramid_1=$derived(constructors[1])
 </script>
 
 {#if constructors[1]}
-	<svelte:component this={constructors[0]} bind:this={components[0]} data={data_0} params={page.params}>
-						<svelte:component this={constructors[1]} bind:this={components[1]} data={data_1} {form} params={page.params} />
-					</svelte:component>
+	{@const Pyramid_0 = constructors[0]}
+							<!-- svelte-ignore binding_property_non_reactive -->
+							<Pyramid_0 bind:this={components[0]} data={data_0} {form} params={page.params}>
+								<!-- svelte-ignore binding_property_non_reactive -->
+										<Pyramid_1 bind:this={components[1]} data={data_1} {form} params={page.params} />
+							</Pyramid_0>
 
 {:else}
-	<svelte:component this={constructors[0]} bind:this={components[0]} data={data_0} {form} params={page.params} />
+	{@const Pyramid_0 = constructors[0]}
+	<!-- svelte-ignore binding_property_non_reactive -->
+	<Pyramid_0 bind:this={components[0]} data={data_0} {form} params={page.params} />
 
 {/if}
 
