@@ -26,77 +26,84 @@
  * ```
  */
 declare module '$env/static/private' {
-	export const STARSHIP_SHELL: string;
-	export const MANPATH: string;
-	export const NIX_PROFILES: string;
-	export const TERM_PROGRAM: string;
-	export const MYBINS: string;
-	export const NODE: string;
-	export const ANDROID_HOME: string;
-	export const TERM: string;
 	export const SHELL: string;
-	export const HOMEBREW_REPOSITORY: string;
-	export const TMPDIR: string;
-	export const VIMWIKIGRAPH_CONFIG: string;
-	export const GARDEN_HOOKS_DIR: string;
-	export const TERM_PROGRAM_VERSION: string;
-	export const WINDOWID: string;
-	export const OBSIDIAN: string;
-	export const npm_config_local_prefix: string;
-	export const USER: string;
-	export const COMMAND_MODE: string;
-	export const OPENAI_API_KEY: string;
-	export const SSH_AUTH_SOCK: string;
-	export const __CF_USER_TEXT_ENCODING: string;
-	export const npm_execpath: string;
-	export const KNOT_TOKEN: string;
-	export const NTFY_SH: string;
-	export const TMUX: string;
-	export const PATH: string;
-	export const npm_package_json: string;
-	export const LaunchInstanceID: string;
-	export const __CFBundleIdentifier: string;
-	export const PWD: string;
 	export const npm_command: string;
-	export const npm_lifecycle_event: string;
-	export const KITTY_PID: string;
-	export const npm_package_name: string;
-	export const LANG: string;
-	export const CLAUDE_SESSION_ID: string;
-	export const NODE_PATH: string;
-	export const TMUX_PANE: string;
-	export const XPC_FLAGS: string;
-	export const GCC_COLORS: string;
-	export const NIX_SSL_CERT_FILE: string;
-	export const ANTHROPIC_API_KEY: string;
-	export const XPC_SERVICE_NAME: string;
-	export const npm_package_version: string;
-	export const GARDEN_HOOKS_ENABLED: string;
-	export const SHLVL: string;
-	export const MANPAGER: string;
-	export const HOME: string;
-	export const TERMINFO: string;
-	export const is_fzf: string;
-	export const HOMEBREW_PREFIX: string;
-	export const STARSHIP_SESSION_KEY: string;
-	export const LOGNAME: string;
-	export const npm_lifecycle_script: string;
-	export const XDG_DATA_DIRS: string;
-	export const TMUX_PLUGIN_MANAGER_PATH: string;
-	export const HOMEBREW_CASK_OPTS: string;
-	export const BUN_INSTALL: string;
-	export const GARDEN_AGENT_ID: string;
-	export const KITTY_INSTALLATION_DIR: string;
-	export const npm_config_user_agent: string;
-	export const KITTY_WINDOW_ID: string;
-	export const HOMEBREW_CELLAR: string;
-	export const INFOPATH: string;
-	export const OSLogRateLimit: string;
-	export const SECURITYSESSIONID: string;
-	export const is_vim: string;
-	export const npm_node_execpath: string;
-	export const KITTY_PUBLIC_KEY: string;
+	export const COREPACK_ENABLE_AUTO_PIN: string;
+	export const WINDOWID: string;
 	export const COLORTERM: string;
+	export const XPC_FLAGS: string;
+	export const TERM_PROGRAM_VERSION: string;
+	export const TMUX: string;
+	export const GARDEN_AGENT_ID: string;
+	export const NODE: string;
+	export const GARDEN_HOOKS_ENABLED: string;
+	export const __CFBundleIdentifier: string;
+	export const SSH_AUTH_SOCK: string;
+	export const ANTHROPIC_API_KEY: string;
+	export const TMUX_PLUGIN_MANAGER_PATH: string;
+	export const OSLogRateLimit: string;
+	export const OPENAI_API_KEY: string;
+	export const npm_config_local_prefix: string;
+	export const HOMEBREW_PREFIX: string;
+	export const KITTY_PID: string;
+	export const is_vim: string;
+	export const PWD: string;
+	export const NIX_PROFILES: string;
+	export const LOGNAME: string;
+	export const MANPATH: string;
+	export const LaunchInstanceID: string;
+	export const _: string;
+	export const KITTY_PUBLIC_KEY: string;
+	export const NoDefaultCurrentDirectoryInExePath: string;
+	export const KNOT_TOKEN: string;
+	export const CLAUDECODE: string;
+	export const COMMAND_MODE: string;
+	export const HOME: string;
+	export const LANG: string;
+	export const npm_package_version: string;
+	export const SECURITYSESSIONID: string;
+	export const STARSHIP_SHELL: string;
+	export const NIX_SSL_CERT_FILE: string;
+	export const KITTY_WINDOW_ID: string;
+	export const TMPDIR: string;
+	export const NTFY_SH: string;
+	export const MYBINS: string;
+	export const CLAUDE_SESSION_ID: string;
+	export const STARSHIP_SESSION_KEY: string;
+	export const INFOPATH: string;
+	export const npm_lifecycle_script: string;
+	export const ANDROID_HOME: string;
+	export const TERMINFO: string;
+	export const TERM: string;
+	export const npm_package_name: string;
+	export const USER: string;
+	export const TMUX_PANE: string;
+	export const HOMEBREW_CELLAR: string;
+	export const MANPAGER: string;
+	export const npm_lifecycle_event: string;
+	export const SHLVL: string;
+	export const GIT_EDITOR: string;
+	export const HOMEBREW_REPOSITORY: string;
+	export const XPC_SERVICE_NAME: string;
+	export const npm_config_user_agent: string;
+	export const OTEL_EXPORTER_OTLP_METRICS_TEMPORALITY_PREFERENCE: string;
+	export const npm_execpath: string;
+	export const GARDEN_HOOKS_DIR: string;
+	export const CLAUDE_CODE_ENTRYPOINT: string;
+	export const npm_package_json: string;
+	export const BUN_INSTALL: string;
+	export const OBSIDIAN: string;
+	export const GCC_COLORS: string;
+	export const XDG_DATA_DIRS: string;
+	export const PATH: string;
+	export const HOMEBREW_CASK_OPTS: string;
+	export const KITTY_INSTALLATION_DIR: string;
+	export const npm_node_execpath: string;
+	export const is_fzf: string;
+	export const OLDPWD: string;
+	export const VIMWIKIGRAPH_CONFIG: string;
+	export const __CF_USER_TEXT_ENCODING: string;
+	export const TERM_PROGRAM: string;
 	export const NODE_ENV: string;
 }
 
@@ -127,77 +134,84 @@ declare module '$env/static/public' {
  */
 declare module '$env/dynamic/private' {
 	export const env: {
-		STARSHIP_SHELL: string;
-		MANPATH: string;
-		NIX_PROFILES: string;
-		TERM_PROGRAM: string;
-		MYBINS: string;
-		NODE: string;
-		ANDROID_HOME: string;
-		TERM: string;
 		SHELL: string;
-		HOMEBREW_REPOSITORY: string;
-		TMPDIR: string;
-		VIMWIKIGRAPH_CONFIG: string;
-		GARDEN_HOOKS_DIR: string;
-		TERM_PROGRAM_VERSION: string;
-		WINDOWID: string;
-		OBSIDIAN: string;
-		npm_config_local_prefix: string;
-		USER: string;
-		COMMAND_MODE: string;
-		OPENAI_API_KEY: string;
-		SSH_AUTH_SOCK: string;
-		__CF_USER_TEXT_ENCODING: string;
-		npm_execpath: string;
-		KNOT_TOKEN: string;
-		NTFY_SH: string;
-		TMUX: string;
-		PATH: string;
-		npm_package_json: string;
-		LaunchInstanceID: string;
-		__CFBundleIdentifier: string;
-		PWD: string;
 		npm_command: string;
-		npm_lifecycle_event: string;
-		KITTY_PID: string;
-		npm_package_name: string;
-		LANG: string;
-		CLAUDE_SESSION_ID: string;
-		NODE_PATH: string;
-		TMUX_PANE: string;
-		XPC_FLAGS: string;
-		GCC_COLORS: string;
-		NIX_SSL_CERT_FILE: string;
-		ANTHROPIC_API_KEY: string;
-		XPC_SERVICE_NAME: string;
-		npm_package_version: string;
-		GARDEN_HOOKS_ENABLED: string;
-		SHLVL: string;
-		MANPAGER: string;
-		HOME: string;
-		TERMINFO: string;
-		is_fzf: string;
-		HOMEBREW_PREFIX: string;
-		STARSHIP_SESSION_KEY: string;
-		LOGNAME: string;
-		npm_lifecycle_script: string;
-		XDG_DATA_DIRS: string;
-		TMUX_PLUGIN_MANAGER_PATH: string;
-		HOMEBREW_CASK_OPTS: string;
-		BUN_INSTALL: string;
-		GARDEN_AGENT_ID: string;
-		KITTY_INSTALLATION_DIR: string;
-		npm_config_user_agent: string;
-		KITTY_WINDOW_ID: string;
-		HOMEBREW_CELLAR: string;
-		INFOPATH: string;
-		OSLogRateLimit: string;
-		SECURITYSESSIONID: string;
-		is_vim: string;
-		npm_node_execpath: string;
-		KITTY_PUBLIC_KEY: string;
+		COREPACK_ENABLE_AUTO_PIN: string;
+		WINDOWID: string;
 		COLORTERM: string;
+		XPC_FLAGS: string;
+		TERM_PROGRAM_VERSION: string;
+		TMUX: string;
+		GARDEN_AGENT_ID: string;
+		NODE: string;
+		GARDEN_HOOKS_ENABLED: string;
+		__CFBundleIdentifier: string;
+		SSH_AUTH_SOCK: string;
+		ANTHROPIC_API_KEY: string;
+		TMUX_PLUGIN_MANAGER_PATH: string;
+		OSLogRateLimit: string;
+		OPENAI_API_KEY: string;
+		npm_config_local_prefix: string;
+		HOMEBREW_PREFIX: string;
+		KITTY_PID: string;
+		is_vim: string;
+		PWD: string;
+		NIX_PROFILES: string;
+		LOGNAME: string;
+		MANPATH: string;
+		LaunchInstanceID: string;
+		_: string;
+		KITTY_PUBLIC_KEY: string;
+		NoDefaultCurrentDirectoryInExePath: string;
+		KNOT_TOKEN: string;
+		CLAUDECODE: string;
+		COMMAND_MODE: string;
+		HOME: string;
+		LANG: string;
+		npm_package_version: string;
+		SECURITYSESSIONID: string;
+		STARSHIP_SHELL: string;
+		NIX_SSL_CERT_FILE: string;
+		KITTY_WINDOW_ID: string;
+		TMPDIR: string;
+		NTFY_SH: string;
+		MYBINS: string;
+		CLAUDE_SESSION_ID: string;
+		STARSHIP_SESSION_KEY: string;
+		INFOPATH: string;
+		npm_lifecycle_script: string;
+		ANDROID_HOME: string;
+		TERMINFO: string;
+		TERM: string;
+		npm_package_name: string;
+		USER: string;
+		TMUX_PANE: string;
+		HOMEBREW_CELLAR: string;
+		MANPAGER: string;
+		npm_lifecycle_event: string;
+		SHLVL: string;
+		GIT_EDITOR: string;
+		HOMEBREW_REPOSITORY: string;
+		XPC_SERVICE_NAME: string;
+		npm_config_user_agent: string;
+		OTEL_EXPORTER_OTLP_METRICS_TEMPORALITY_PREFERENCE: string;
+		npm_execpath: string;
+		GARDEN_HOOKS_DIR: string;
+		CLAUDE_CODE_ENTRYPOINT: string;
+		npm_package_json: string;
+		BUN_INSTALL: string;
+		OBSIDIAN: string;
+		GCC_COLORS: string;
+		XDG_DATA_DIRS: string;
+		PATH: string;
+		HOMEBREW_CASK_OPTS: string;
+		KITTY_INSTALLATION_DIR: string;
+		npm_node_execpath: string;
+		is_fzf: string;
+		OLDPWD: string;
+		VIMWIKIGRAPH_CONFIG: string;
+		__CF_USER_TEXT_ENCODING: string;
+		TERM_PROGRAM: string;
 		NODE_ENV: string;
 		[key: `PUBLIC_${string}`]: undefined;
 		[key: `${string}`]: string | undefined;
