@@ -1,1 +1,1 @@
-export { default as component } from "../../../../node_modules/.pnpm/@sveltejs+kit@2.37.0_@sveltejs+vite-plugin-svelte@3.1.2_svelte@4.2.20_vite@5.4.19__svelte@4.2.20_vite@5.4.19/node_modules/@sveltejs/kit/src/runtime/components/svelte-4/error.svelte";
+export { default as component } from "../../../../node_modules/@sveltejs/kit/src/runtime/components/svelte-5/error.svelte";
